move tool caller definition out of agent module
